assignments/workhuman/requirements.txt
pandas
pyarrow
requests
PyYAML
pytest
